bot_Eva/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bot-eva"
version = "1.0.0"
description = "Sophisticated Binance trading bot with trend following strategy and adaptive risk management"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "DZX-CORE", email = "contact@dzx-core.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
keywords = ["trading", "binance", "cryptocurrency", "algorithmic-trading", "bot", "eva"]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.11.0",
    "python-binance>=1.0.19",
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "python-dateutil>=2.8.2",
    "aiohttp>=3.8.0",
    "websockets>=11.0.0",
    "pycryptodome>=3.15.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/DZX-CORE/bot_Eva"
Repository = "https://github.com/DZX-CORE/bot_Eva.git"
Issues = "https://github.com/DZX-CORE/bot_Eva/issues"

[project.scripts]
trading-bot = "main:main"
demo-bot = "demo_bot:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true